fix(admin): theme the puck editor, fix overlays, and modernize controls
- Map Puck's color ramps to the active theme so editor chrome (fields, labels,
  the tiptap toolbar, sidebars) follows dark/light; keep the ActionBar a dark chip.
- Give Puck's block wrappers a block-formatting context (flow-root) so selection
  and hover overlays include block margins and align with the visible block.
- Thin, theme-aware scrollbars across the editor surfaces.
- Toast on Copy JSON / Copy as pattern source so the action gives feedback.
- Replace the bare copy-preview icon with a Preview split-button that opens the
  signed draft preview in a new tab; copy-link moves into its menu.
- Swap native data-table checkboxes and token-scope radios for shadcn Checkbox
  and RadioGroup.
- Consolidate the editor e2e into one feature-grouped editor.spec.ts (69 tests).
