+++ packages/admin/src/components/profile/api-tokens-card.tsx
@@ -37,6 +37,7 @@ import {
 } from "@/components/ui/form.js";
 import { Input } from "@/components/ui/input.js";
 import { Label as UILabel } from "@/components/ui/label.js";
+import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group.js";
 import {
   Table,
   TableBody,
@@ -606,38 +607,46 @@ function CreateTokenForm({
                 />
               </FormLabel>
               <FormControl>
-                <div className="flex flex-col gap-2">
-                  <UILabel className="flex items-center gap-2 font-normal">
-                    <input
-                      type="radio"
-                      name={field.name}
+                <RadioGroup
+                  value={field.value}
+                  onValueChange={field.onChange}
+                  className="gap-2"
+                >
+                  <div className="flex items-center gap-2">
+                    <RadioGroupItem
                       value="inherit"
-                      checked={field.value === "inherit"}
-                      onChange={() => field.onChange("inherit")}
+                      id="api-token-scope-inherit"
                       disabled={pending}
                       data-testid="api-tokens-create-scope-inherit-radio"
                     />
-                    <Trans
-                      id="apiTokens.create.scope.inherit"
-                      message="Inherit all your permissions"
-                    />
-                  </UILabel>
-                  <UILabel className="flex items-center gap-2 font-normal">
-                    <input
-                      type="radio"
-                      name={field.name}
+                    <UILabel
+                      htmlFor="api-token-scope-inherit"
+                      className="font-normal"
+                    >
+                      <Trans
+                        id="apiTokens.create.scope.inherit"
+                        message="Inherit all your permissions"
+                      />
+                    </UILabel>
+                  </div>
+                  <div className="flex items-center gap-2">
+                    <RadioGroupItem
                       value="restrict"
-                      checked={field.value === "restrict"}
-                      onChange={() => field.onChange("restrict")}
+                      id="api-token-scope-restrict"
                       disabled={pending}
                       data-testid="api-tokens-create-scope-restrict-radio"
                     />
-                    <Trans
-                      id="apiTokens.create.scope.restrict"
-                      message="Restrict to specific capabilities"
-                    />
-                  </UILabel>
-                </div>
+                    <UILabel
+                      htmlFor="api-token-scope-restrict"
+                      className="font-normal"
+                    >
+                      <Trans
+                        id="apiTokens.create.scope.restrict"
+                        message="Restrict to specific capabilities"
+                      />
+                    </UILabel>
+                  </div>
+                </RadioGroup>
               </FormControl>
               <FormMessage />
             </FormItem>
